backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
Pillow==10.4.0

# Scraper dependencies
anthropic==0.21.0
beautifulsoup4==4.12.0
html5lib==1.1
requests==2.31.0
aiohttp==3.9.0

# Required for auth, hashing, JWT, multipart, and DB drivers
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.6
python-multipart==0.0.6
typing-extensions==4.8.0

requests  # For password breach checking
redis  # Optional, for distributed rate limiting

bleach==6.1.0
psutil==5.9.6

requests==2.31.0  # For currency API calls

boto3>=1.34.0

apscheduler==3.10.4
